feat: allow to override HMAC validation
moved logic from HMAC guard into a service that can be overridden
