+++ .changeset/thin-mice-type.md
@@ -1,0 +1,6 @@
+---
+"@nestjs-shopify/webhooks": minor
+"@nestjs-shopify/core": minor
+---
+
+allow to override HMAC validation
