+++ packages/core/src/hmac/hmac.module.ts
@@ -1,8 +1,10 @@
 import { Global, Module } from '@nestjs/common';
 import { ShopifyHmacGuard } from './hmac.guard';
+import { ShopifyHmacService } from './hmac.service';
 
 @Global()
 @Module({
-  providers: [ShopifyHmacGuard],
+  providers: [ShopifyHmacGuard, ShopifyHmacService],
+  exports: [ShopifyHmacService],
 })
 export class ShopifyHmacModule {}
