+++ packages/core/src/hmac/index.ts
@@ -1,4 +1,5 @@
 export * from './hmac.decorators';
 export * from './hmac.enums';
+export * from './hmac.service';
 export * from './hmac.guard';
 export * from './hmac.module';
